Create q1.py
Visualization 1. A visualization of speeds for a single trip for any bus that crossed the Tilikum Crossing Bridge. You choose the day, time and route for your selected trip. To find a trip that traverses this bridge, consider finding a trip that includes breadcrumbs within the rectangle bounded by (45.505621, -122.668632),  and (45.504416, -122.665738).